refactor: separate out alignment guides
