+++ frontend/src/components/Slide.vue
@@ -17,20 +17,7 @@
 			:class="activeElement?.type == 'slide' ? 'ring-[1px] ring-gray-200' : ''"
 			@click="handleSlideClick"
 		>
-			<div
-				v-show="isDragging && showVerticalCenter"
-				class="absolute left-1/2 h-full w-[1px] -translate-x-1/2 bg-blue-400"
-			></div>
-
-			<div
-				v-show="isDragging && showHorizontalCenter"
-				class="absolute top-1/2 h-[1px] w-full -translate-y-1/2 bg-blue-400"
-			></div>
-
-			<div v-show="diffLeft < 10" :style="leftGuideStyles"></div>
-			<div v-show="diffRight < 10" :style="rightGuideStyles"></div>
-			<div v-show="diffTop < 10" :style="topGuideStyles"></div>
-			<div v-show="diffBottom < 10" :style="bottomGuideStyles"></div>
+			<ElementAlignmentGuides v-if="isDragging" :slideRect="slideRect" />
 
 			<div v-if="activeSlideElements">
 				<TransitionGroup
@@ -54,7 +41,6 @@
 					:is="SlideElement"
 					:element="element"
 					:data-index="index"
-					:currentPairedDataIndex="currentPairedDataIndex"
 				/>
 			</div>
 		</div>
@@ -76,6 +62,7 @@ import {
 import { useElementBounding } from '@vueuse/core'
 
 import SlideElement from '@/components/SlideElement.vue'
+import ElementAlignmentGuides from '@/components/ElementAlignmentGuides.vue'
 
 import { useDragAndDrop } from '@/utils/drag'
 import { useResizer } from '@/utils/resizer'
@@ -88,6 +75,8 @@ import {
 	presentation,
 	activeSlideElements,
 	inSlideShow,
+	position,
+	dimensions,
 } from '@/stores/slide'
 import html2canvas from 'html2canvas'
 
@@ -99,8 +88,6 @@ const targetRef = useTemplateRef('target')
 const { transform, transformOrigin, allowPanAndZoom } = zoom.value
 
 const slideCursor = ref('auto')
-const position = ref(null)
-const dimensions = ref(null)
 
 const { isDragging, dragTarget } = useDragAndDrop(position)
 const { isResizing, resizeTarget, resizeMode } = useResizer(position, dimensions)
@@ -295,177 +282,7 @@ watch(
 	{ immediate: true },
 )
 
-const activeDiv = computed(() => {
-	return document.querySelector(`[data-index="${currentDataIndex.value}"]`)
-})
-
-const pairedDiv = computed(() => {
-	return document.querySelector(`[data-index="${currentPairedDataIndex.value}"]`)
-})
-
-const currentPairedDataIndex = computed(() => {
-	if (!activeElement.value) return
-	let i = null
-	activeSlideElements.value.forEach((element, index) => {
-		if (index == currentDataIndex.value) return
-		let diffLeft = Math.abs(element.left - activeElement.value.left)
-		let diffRight = Math.abs(
-			element.left + element.width - activeElement.value.left - activeElement.value.width,
-		)
-		let diffTop = Math.abs(element.top - activeElement.value.top)
-		let elementDiv = document.querySelector(`[data-index="${index}"]`).getBoundingClientRect()
-		let diffBottom = Math.abs(
-			element.top + elementDiv.height - activeElement.value.top - activeRect.height.value,
-		)
-		if (diffLeft < 10 || diffRight < 10 || diffTop < 10 || diffBottom < 10) i = index
-	})
-	return i
-})
-
 const slideRect = useElementBounding(targetRef)
-const activeRect = useElementBounding(activeDiv)
-const pairedRect = useElementBounding(pairedDiv)
-
-const pairElement = computed(() => {
-	if (!isDragging.value) return
-	return activeSlideElements.value[currentPairedDataIndex.value]
-})
-
-const snapToPairElement = () => {
-	if (!pairElement.value) return
-	if (diffLeft.value < 10) {
-		activeElement.value.left = pairElement.value.left
-	}
-	if (diffRight.value < 10) {
-		activeElement.value.left =
-			pairElement.value.left + pairElement.value.width - activeElement.value.width
-	}
-	if (diffTop.value < 10) {
-		activeElement.value.top = pairElement.value.top
-	}
-	if (diffBottom.value < 10) {
-		activeElement.value.top =
-			pairElement.value.top + pairedRect.height.value - activeRect.height.value
-	}
-}
-
-const showVerticalCenter = ref(false)
-const showHorizontalCenter = ref(false)
-
-const diffLeft = computed(() => {
-	if (!pairElement.value) return
-	return Math.abs(pairElement.value.left - activeElement.value.left)
-})
-
-const diffRight = computed(() => {
-	if (!pairElement.value) return
-	return Math.abs(
-		pairElement.value.left +
-			pairElement.value.width -
-			activeElement.value.left -
-			activeElement.value.width,
-	)
-})
-
-const diffTop = computed(() => {
-	if (!pairElement.value) return
-	return Math.abs(pairElement.value.top - activeElement.value.top)
-})
-
-const diffBottom = computed(() => {
-	if (!pairElement.value) return
-	return Math.abs(
-		pairElement.value.top +
-			pairedRect.height.value -
-			activeElement.value.top -
-			activeRect.height.value,
-	)
-})
-
-const guideStyles = {
-	position: 'fixed',
-	borderColor: '#70b6f0',
-	borderStyle: 'dashed',
-}
-
-const leftGuideStyles = computed(() => {
-	if (!pairElement.value) return
-	return {
-		...guideStyles,
-		borderWidth: '0 0 0 1px',
-		left: `${activeElement.value.left - 5}px`,
-		top: `${Math.min(pairElement.value.top, activeElement.value.top)}px`,
-		height: `${Math.abs(pairElement.value.top - activeElement.value.top)}px`,
-	}
-})
-
-const rightGuideStyles = computed(() => {
-	if (!pairElement.value) return
-	return {
-		...guideStyles,
-		borderWidth: '0 0 0 1px',
-		left: `${activeElement.value.left + activeElement.value.width + 4.5}px`,
-		top: `${Math.min(pairElement.value.top, activeElement.value.top)}px`,
-		height: `${Math.abs(pairElement.value.top - activeElement.value.top)}px`,
-	}
-})
-
-const topGuideStyles = computed(() => {
-	if (!pairElement.value) return
-	return {
-		...guideStyles,
-		borderWidth: '1px 0 0 0',
-		top: `${activeElement.value.top - 5}px`,
-		left: `${Math.min(pairElement.value.left, activeElement.value.left)}px`,
-		width: `${Math.abs(pairElement.value.left - activeElement.value.left)}px`,
-	}
-})
-
-const bottomGuideStyles = computed(() => {
-	if (!pairElement.value) return
-	return {
-		...guideStyles,
-		borderWidth: '1px 0 0 0',
-		top: `${activeElement.value.top + activeRect.height.value + 4.5}px`,
-		left: `${Math.min(pairElement.value.left, activeElement.value.left)}px`,
-		width: `${Math.abs(pairElement.value.left - activeElement.value.left)}px`,
-	}
-})
-
-const updateCenterAlignmentGuides = () => {
-	let centerX = slideRect.left.value + slideRect.width.value / 2
-	let centerY = slideRect.top.value + slideRect.height.value / 2
-
-	let centerOfElementX = position.value.left + activeRect.width.value / 2
-	let centerOfElementY = position.value.top + activeRect.height.value / 2
-
-	showVerticalCenter.value = Math.abs(centerOfElementX - centerX) < 10
-	showHorizontalCenter.value = Math.abs(centerOfElementY - centerY) < 10
-}
-
-watch(
-	() => isDragging.value,
-	() => {
-		if (!isDragging.value) {
-			let centerX = slideRect.left.value + slideRect.width.value / 2
-			let centerY = slideRect.top.value + slideRect.height.value / 2
-
-			if (showVerticalCenter.value) {
-				position.value = {
-					...position.value,
-					left: centerX - activeRect.width.value / 2,
-				}
-			}
-
-			if (showHorizontalCenter.value) {
-				position.value = { ...position.value, top: centerY - activeRect.height.value / 2 }
-			}
-
-			showVerticalCenter.value = false
-			showHorizontalCenter.value = false
-		}
-	},
-)
 
 watch(
 	() => position.value,
@@ -474,9 +291,6 @@ watch(
 		let currentScale = slideRect.width.value / 960
 		activeElement.value.left = (position.value.left - slideRect.left.value) / currentScale
 		activeElement.value.top = (position.value.top - slideRect.top.value) / currentScale
-
-		updateCenterAlignmentGuides()
-		snapToPairElement()
 	},
 	{ immediate: true },
 )
